+++ stegonoscreens/src/components/imageHandler/ImageUpload.js
@@ -1,19 +1,18 @@
-import { useState } from 'react';
-import classes from './ImageUpload.module.css';
-
-import ErrorModal from './ErrorModal';
+import { useState, useContext } from 'react';
+import classes from './style/ImageUpload.module.css';
 
 import Form from 'react-bootstrap/Form';
 
-import errors from '../static/errors.js';
-import upload from '../static/icons/upload.svg';
+import AppContext from '../../store/app-context.js'
+import config from '../../configs/config.json'
+import errors from '../../static/errors.js';
+import upload from '../../static/icons/upload.svg';
 
 function ImageUpload(props) {
-    const sizeLimit = 2097152;
+    const sizeLimit = config.imageSizeLimit;
 
+    const appCtx = useContext(AppContext);
     const [displayedImage, setDisplayedImage] = useState(upload);
-    const [error, setError] = useState(null);
-    const [showError, setShowError] = useState(false);
 
     function submitHandler(event, func){
         const file = event.target.files[0]
@@ -25,18 +24,13 @@ function ImageUpload(props) {
                 setDisplayedImage(fileForDisplay);
             }
             else{
-                setError(errors.imgTooLarge)
-                setShowError(true);
+                appCtx.raiseError(errors.imgTooLarge)
+                appCtx.setShowError(true);
             };
         } else{
             func(null);
             setDisplayedImage(upload);
         }
-    };
-
-    function closeErrorModal(){
-        setError(null);
-        setShowError(false);
     };
 
     return <div className={classes.container}>
@@ -47,7 +41,6 @@ function ImageUpload(props) {
         <Form.Control type='file' accept='image/png' size='sm' 
             onChange={(e) => submitHandler(e, props.func)}>
         </Form.Control>
-        <ErrorModal showModal={showError} showHandler={closeErrorModal} error={error}/>
     </div>;
 };
