+++ stegonoscreens/src/components/misc/style/Question.module.css
