+++ stegonoscreens/src/components/modals/LoadingModal.js
@@ -3,9 +3,9 @@ import { useContext } from 'react';
 import Spinner from 'react-bootstrap/Spinner';
 import Modal from 'react-bootstrap/Modal';
 
-import AppContext from '../store/app-context.js'
+import AppContext from '../../store/app-context.js'
 
-import classes from './LoadingModal.module.css';
+import classes from './style/LoadingModal.module.css';
 
 function LoadingModal(props) {
     const appCtx = useContext(AppContext);
