+++ stegonoscreens/src/components/modals/style/ResultModal.module.css
